+++ examples/mobile_omni_wheeled_robot.py
@@ -64,8 +64,8 @@ if __name__ == "__main__":
 
     # Select QP solver
     solver = qpsolvers.available_solvers[0]
-    if "quadprog" in qpsolvers.available_solvers:
-        solver = "quadprog"
+    if "daqp" in qpsolvers.available_solvers:
+        solver = "daqp"
 
     rate = RateLimiter(frequency=100.0, warn=False)
     dt = rate.period