+++ pink/__init__.py
@@ -19,7 +19,7 @@ from .tasks import (
 )
 from .utils import custom_configuration_vector
 
-__version__ = "3.1.0"
+__version__ = "3.2.0"
 
 __all__ = [
     "Configuration",