+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Update supported Python versions to 3.9–3.12
+
 ## [3.1.0] - 2024-10-28
 
 ### Added