+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CICD: Update checkout action to v4
 - Update supported Python versions to 3.9–3.12
 
+### Fixed
+
+- CICD: Update unit test for self-collision barrier
+
 ## [3.1.0] - 2024-10-28
 
 ### Added