+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - CICD: Update unit test for self-collision barrier
 
+### Removed
+
+- Remove dependency on quadprog (thanks to @peterd-NV)
+
 ## [3.1.0] - 2024-10-28
 
 ### Added