+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.2.0] - 2025-04-22
+
 ### Added
 
 - Add `NoSolutionFound` exception
@@ -23,7 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CICD: Update QP solvers used in unit tests
 - CICD: Update checkout action to v4
 - Raise custom `NoSolutionFound` rather than `AssertionError` in `solve_ik`
-- Update supported Python versions to 3.9–3.12
+- Update supported Python versions
 
 ### Fixed
 
@@ -357,7 +359,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Python package infrastructure
 
-[unreleased]: https://github.com/qpsolvers/qpsolvers/compare/v3.1.0...HEAD
+[unreleased]: https://github.com/qpsolvers/qpsolvers/compare/v3.2.0...HEAD
+[3.2.0]: https://github.com/qpsolvers/qpsolvers/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/qpsolvers/qpsolvers/compare/v3.0.0...v3.1.0
 [3.0.0]: https://github.com/qpsolvers/qpsolvers/compare/v2.1.0...v3.0.0
 [2.1.0]: https://github.com/qpsolvers/qpsolvers/compare/v2.0.0...v2.1.0